@@ -1,6 +1,6 @@
 {
   "name": "appc-pubsub",
-  "version": "1.4.2",
+  "version": "1.5.0",
   "description": "AppC pubsub client library",
   "main": "index.js",
   "scripts": {
@@ -21,15 +21,19 @@
   },
   "homepage": "https://github.com/appcelerator/appc-pubsub",
   "dependencies": {
-    "basic-auth": "^2.0.0",
-    "request": "^2.88.0"
+    "basic-auth": "2.0.1",
+    "request": "2.88.2"
   },
   "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-config-axway": "^2.0.14",
-    "eslint-plugin-mocha": "^5.0.0",
-    "mocha": "^5.2.0",
-    "nyc": "^12.0.2",
-    "should": "^13.2.1"
+    "eslint": "7.31.0",
+    "eslint-config-axway": "6.0.2",
+    "eslint-plugin-mocha": "9.0.0",
+    "mocha": "9.0.3",
+    "nyc": "15.1.0",
+    "should": "13.2.3"
+  },
+  "engines": {
+    "node": "^12.22.3",
+    "npm": "^6.12.0"
   }
 }
