@@ -2,7 +2,6 @@
 
 const basicAuth = require('basic-auth');
 const exec = require('child_process').exec;
-const url = require('url');
 const crypto = require('crypto');
 const util = require('util');
 const request = require('request');
@@ -154,9 +153,10 @@ util.inherits(PubSubClient, EventEmitter);
  */
 PubSubClient.prototype.fetchConfig = function () {
 	// some random data to sign for the signature
+	const url = new URL('/api/client/config', this.url);
 	let data = {},
 		opts = {
-			url: url.resolve(this.url, '/api/client/config'),
+			url: url.href,
 			method: 'get',
 			json: data,
 			headers: {
@@ -203,9 +203,9 @@ PubSubClient.prototype._parseConfig = function (data) {
 
 		// Get basic auth creds from the url
 		if (data.auth_type === 'basic' && data.url) {
-			let details = (url.parse(data.url) || '').auth.split(':');
-			data.auth_user = details[0];
-			data.auth_pass = details[1];
+			const url = new URL(data.url);
+			data.auth_user = url.username;
+			data.auth_pass = url.password;
 		}
 	}
 	this.config = data;
@@ -347,7 +347,7 @@ PubSubClient.prototype.getFingerprint = function (callback) {
 };
 
 function serialize(obj, seen) {
-	if (!obj || typeof(obj) !== 'object') {
+	if (!obj || typeof obj !== 'object') {
 		return obj;
 	}
 	if (obj instanceof RegExp) {
@@ -358,7 +358,7 @@ function serialize(obj, seen) {
 	}
 	Object.keys(obj).forEach(function (key) {
 		var value = obj[key],
-			t = typeof(value);
+			t = typeof value;
 		if (t === 'function') {
 			delete obj[key];
 		} else if (/^(password|creditcard)/.test(key)) {
@@ -399,7 +399,7 @@ PubSubClient.prototype.publish = function (name, data, options) {
 	if (Buffer.byteLength(name) > 255) {
 		throw new Error('name length must be less than 255 bytes');
 	}
-	if (data && typeof(data) !== 'object') {
+	if (data && typeof data !== 'object') {
 		throw new Error('data must be an object');
 	}
 	// Clone data before serialization pass so objects are not modified.
@@ -464,8 +464,9 @@ PubSubClient.prototype._send = function (data) {
 	let ticket = Date.now();
 	this._sendingTS = ticket;
 
+	const url = new URL('/api/event', this.url);
 	let opts = {
-		url: url.resolve(this.url, '/api/event'),
+		url: url.href,
 		method: 'post',
 		json: data,
 		headers: {
